+++ packages/containers/src/Notification/Notification.test.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { Provider } from 'react-cmf/lib/mock';
-import { Map } from 'immutable';
-
+import { store, Provider } from 'react-cmf/lib/mock';
+import { fromJS, Map } from 'immutable';
 import Container, { DEFAULT_STATE } from './Notification.container';
 import Connected, {
 	mapDispatchToProps,
 	mapStateToProps,
 } from './Notification.connect';
+import pushNotification from './pushNotification';
 
 describe('Container Notification', () => {
 	it('should render', () => {
@@ -45,3 +45,21 @@ describe('Connected Notification', () => {
 	});
 });
 
+describe('Notification.pushNotification', () => {
+	it('should add a Notification in the state', () => {
+		const state = store.state();
+		state.cmf.components = fromJS({
+			Notification: {
+				Notification: {
+					notifications: [],
+				},
+			},
+		});
+		const notification = { message: 'hello world' };
+		const newState = pushNotification(state, notification);
+		expect(newState).not.toBe(state);
+		const notifications = newState.cmf.components.getIn(['Notification', 'Notification', 'notifications']);
+		expect(notifications.size).toBe(1);
+		expect(notifications.get(0).message).toBe('hello world');
+	});
+});
