+++ packages/containers/src/Notification/index.js
@@ -1,15 +1,5 @@
-import objectId from 'bson-objectid';
 import Notification from './Notification.connect';
+import pushNotification from './pushNotification';
 
-Notification.push = function pushNotification(state, notification) {
-	const path = ['Notification', 'Notification', 'notifications'];
-	let notifs = state.cmf.components.getIn(path);
-	notifs = notifs.push(Object.assign({
-		id: objectId(),
-	}, notification));
-	const newState = Object.assign({}, state);
-	newState.cmf.components = state.cmf.components.setIn(path, notifs);
-	return newState;
-};
-
+Notification.push = pushNotification;
 export default Notification;
